Fix type on amounts. General exception catch

+++ src/main/groovy/com/intuit/workshop/invoicing/domain/entity/Invoice.groovy
@@ -17,8 +17,7 @@ class Invoice implements Entity {
 
     Boolean paid
 
-    // TODO Amount
-    Integer totalAmount
+    BigDecimal totalAmount
 
     User user
